Worked on GUI
The basic gui has been made although not functional as of yet. The gui python file has been made capable of interacting with the main file.
Version 3.0.0.1

+++ Reminder/reminder.py
@@ -4,13 +4,21 @@ from time import sleep
 import custom_reminder
 import getpass
 import os
+import reminder_gui
 
 def add_to_startup():
-    with open(r"C:\ProgramData\216c40503e\reminder.bat","w") as file:
-        file_dir = os.path.dirname(os.path.realpath(__file__))
-        file_name = "\\reminder.exe"
-        file_location = file_dir + file_name
-        file.write("Start %s "%file_location )
+    if os.path.exists(r"C:\ProgramData\216c40503e"):
+        with open(r"C:\ProgramData\216c40503e\reminder.bat","w") as file:
+            file_dir = os.path.dirname(os.path.realpath(__file__))
+            file_name = "\\reminder.exe"
+            file_location = file_dir + file_name
+            file.write("Start %s "%file_location)
+    else:
+        with open(r"C:\Users\USER\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup\reminder.bat", "w") as file:
+            file_dir = os.path.dirname(os.path.realpath(__file__))
+            file_name = "\\reminder.exe"
+            file_location = file_dir + file_name
+            file.write("Start %s " % file_location)
 
 last_reminder_sent = int(datetime.datetime.now().strftime("%M"))
 def is_time():#checks if it is time yet for the notification to be sent
@@ -42,7 +50,9 @@ while True:
             notification.notify(title = "Take a break",
                                 message = "Take a break, rest your eyes, move around and drink water",
                                 timeout = 30,
-                                app_icon = "remindericon.ico")
+                                app_icon = "icons\\remindericon.ico")
             last_reminder_sent = int(datetime.datetime.now().strftime("%M"))
         else:
             sleep(60)
+#
+# reminder_gui.submit()
